[GLUTEN-12181][CH] Fix issues for the Spark 3.5 + Delta Delete Vector with the CH backend
Fix issues for the Spark 3.5 + Delta Delete Vector with the CH backend: enable the suite GlutenDeltaParquetDeletionVectorSuite
