+++ backends-clickhouse/src/test/scala/org/apache/gluten/execution/hive/GlutenClickHouseNativeWriteTableSuite.scala
@@ -327,7 +327,7 @@ class GlutenClickHouseNativeWriteTableSuite
     }
   }
 
-  // TODO: after rebase-25.12, failed with spark35 (Memory limit exceeded), fix later
+  // TODO: after rebase-25.12, failed with spark35 + Orc (Memory limit exceeded), fix later
   testWithSpecifiedSparkVersion("test 2-col partitioned table", "3.3") {
     val fields: ListMap[String, String] = ListMap(
       ("string_field", "string"),
@@ -551,6 +551,7 @@ class GlutenClickHouseNativeWriteTableSuite
     }
   }
 
+  // TODO: after rebase-25.12, failed with spark35 + Orc (Memory limit exceeded), fix later
   testWithMaxSparkVersion("test 1-col partitioned + 2-col bucketed table", "3.3") {
     val fields: ListMap[String, String] = ListMap(
       ("string_field", "string"),
@@ -625,6 +626,7 @@ class GlutenClickHouseNativeWriteTableSuite
     }
   }
 
+  // TODO: after rebase-25.12, failed with spark35 + Orc (Memory limit exceeded), fix later
   testWithMaxSparkVersion("test decimal with rand()", "3.3") {
     nativeWrite {
       format =>
