+++ cpp-ch/local-engine/Storages/SubstraitSource/Delta/DeltaWriter.cpp
@@ -190,7 +190,7 @@ void DeltaWriter::writeDeletionVector(const DB::Block & block)
                 {
                     // use already existing deletion vector
                     auto dv_descriptor_field = existing_deletion_vector_descriptor.createDeletionVectorDescriptorField();
-                    file_path_column->insert(file_path.data());
+                    file_path_column->insert(file_path);
                     dv_descriptor_column->insert(dv_descriptor_field);
                     matched_row_count_col->insert(cardinality);
                     continue;
@@ -214,7 +214,7 @@ void DeltaWriter::writeDeletionVector(const DB::Block & block)
         DeletionVectorDescriptor descriptor(cardinality, bitmap_size, DeletionVectorDescriptor::UUID_DV_MARKER, offset, prefix + encoded);
         auto dv_descriptor_field = descriptor.createDeletionVectorDescriptorField();
 
-        file_path_column->insert(file_path.data());
+        file_path_column->insert(file_path);
         dv_descriptor_column->insert(dv_descriptor_field);
         matched_row_count_col->insert(cardinality);
